+++ src/models/dao/database_error.js
