+++ src/routes/videos.js
@@ -1,7 +1,7 @@
 const router = require('express').Router()
 const { statSync } = require('fs')
-const dao = require('../dao/video')
-const { handleError } = require('../dao/database_error')
+const dao = require('../models/dao/video')
+const { handleError } = require('../models/dao/database_error')
 const fs = require("fs")
 const moment = require('moment')
 const tryCatch = require('../controllers/tryCatch')
